+++ .ci/install-x86_64.sh
@@ -3,4 +3,6 @@
 wd=$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 $wd/install-common.sh
 
-${SUDO:-} apt install -yq python3.5-dev
+if ! dpkg -l python3.5-dev; then
+  ${SUDO:-} apt install -yq python3.5-dev
+fi
